reworked permissions test and changed variable names
