+++ components/MultiCropSelector/MultiCropSelector.permissions.comp.cy.js
@@ -12,6 +12,26 @@ describe('Test the MultiCropSelector permissions', () => {
     cy.saveSessionStorage();
   });
 
+  it('Admin can add crops', () => {
+    const readySpy = cy.spy().as('readySpy');
+
+    cy.wrap(
+      farmosUtil.getFarmOSInstance('http://farmos', 'farm', 'admin', 'admin')
+    ).then(() => {
+      cy.mount(MultiCropSelector, {
+        props: {
+          onReady: readySpy,
+        },
+      });
+
+      cy.get('@readySpy')
+        .should('have.been.calledOnce')
+        .then(() => {
+          cy.get('[data-cy="selector-add-button"]').should('exist');
+        });
+    });
+  });
+
   it('Guest cannot add crop', () => {
     const readySpy = cy.spy().as('readySpy');
